This is a temporary, backup commit which probably doesn't even build
This is a backup of all the work I forgot to commit for a long time
after my latest commit - I have not had the time to make sure it even
builds or anything, this is just a backup in case my laptop implodes
or something.

The next commit ought to be an amend commit - if I forget I should
probably rewrite history to eliminate this commit, really given how
likely it is to be completely broken.