fix: Some CMakeLists refactors, TODO and internal scripts updates
This fixes a stupid bug in the CMakeLists.txt to avoid having all of
the YALIBCT_DISABLE_ flags get enabled by default. This also removes
the TODO entry corresponding to the previous commit, and updates an
internal build script for cosmopolitan's latest release to use the
latest release (released yesterday).